Assignment: SQL Select and Aggregations
1. Which destination in the flights database is the furthest distance
away, based on information in the flights table.
Show the SQL query(s) that support your conclusion.
2. What are the different numbers of engines in the planes table? For
each number of engines, which aircraft have
the most number of seats? Show the SQL statement(s) that support your
result.
3. Show the total number of flights.
4. Show the total number of flights by airline (carrier).
5. Show all of the airlines, ordered by number of flights in descending
order.
6. Show only the top 5 airlines, by number of flights, ordered by number
of flights in descending order.
7. Show only the top 5 airlines, by number of flights of distance 1,000
miles or greater, ordered by number of
flights in descending order.
8. Create a question that (a) uses data from the flights database, and
(b) requires aggregation to answer it, and
write down both the question, and the query that answers the question.